[FIX] put the scrollbar back in web/tests
for some reason, the web tests interface loads base.css, in which the .openerp is overflow:hidden, which is not fun in the tests.

The correct fix is actually to remove base.css from the tests, but the way I did it had some problems. So, this is a simple workaround that can be safely removed if base.css is removed from the view.