[IMP] hr_holidays: when searching on holidays.status with an employee_id in context and no specific order, order on limited leaves then on remaining leaves. This is usefull for users having their normal leaves being displayed in top of the list (legal leaves, compensatory leaves, then sick and unpaid for example).
Also updated the name_get of holidays status, displaying (remaining leaves /
max leaves) instead of (taken leaves / max leaves). What is important for
users is the number of remaining leaves, not the number of leaves already
taken in that type.

Also updated and reindented demo data (compensatory is now limited + order
of declaration due to this change).