[FIX] crm: do not assign a salesteam to a lead if not in multi sales team env
on_change_user was used to assign the first team  in which the user is a member to the lead
When the user does not use the multi sales team, it therefore set a default sales team, but invisible to the user.

Stages displayed in the kanban view are the lead sales team stages. In a non sales team env, only stages with no stages are displayed for new leads
If you added a new stage, in the kanban view, the stage is not assigned to a team